Refactoring

+++ DistributionOfStudents/Controllers/AdmissionsController.cs
@@ -1,18 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using DistributionOfStudents.Data;
+﻿using DistributionOfStudents.Data.Interfaces;
 using DistributionOfStudents.Data.Models;
-using DistributionOfStudents.Data.Interfaces;
 using DistributionOfStudents.Data.Specifications;
-using DistributionOfStudents.Data.Repositories;
-using Microsoft.Extensions.Logging;
-using NuGet.Protocol.Plugins;
-using DistributionOfStudents.ViewModels;
+using DistributionOfStudents.ViewModels.Admissions;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 
 namespace DistributionOfStudents.Controllers
 {
@@ -38,58 +29,33 @@ namespace DistributionOfStudents.Controllers
         // GET: Admissions/Details/5
         public async Task<IActionResult> Details(string facultyName, int groupId, int id)
         {
-           
-            Admission admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
+            Admission? admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
 
             if (admission == null)
             {
                 return NotFound();
             }
             admission.SpecialityPriorities = admission.SpecialityPriorities.OrderBy(i => i.Priority).ToList();
-            DetailsAdmissionVM model = new()
-            {
-                FacultyName = facultyName,
-                Admission= admission,
-            };
 
-            return View(model);
+            return View(admission);
         }
 
         // GET: Admissions/Create
         public async Task<IActionResult> Create(string facultyName, int groupId)
         {
-            GroupOfSpecialties group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
             List<StudentScore> scores = new();
             List<SpecialityPriorityVM> priorities = new();
+            GroupOfSpecialties? group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
+            if (group == null)
+            {
+                return NotFound();
+            }
             List<RecruitmentPlan> plans = await _planRepository.GetAllAsync(new RecruitmentPlansSpecification().WhereFaculty(facultyName).WhereGroup(group));
-            plans = plans.Where(p => group.Specialities.Contains(p.Speciality)).ToList();
+            plans = plans.Where(p => (group.Specialities ?? new()).Contains(p.Speciality)).ToList();
+            (group.Subjects ?? new()).ForEach(subject => scores.Add(new StudentScore() { Subject = subject }));
+            plans.ForEach(plan => priorities.Add(new SpecialityPriorityVM(plan)));
 
-            foreach (Subject subject in group.Subjects)
-            {
-                StudentScore score = new()
-                {
-                    Subject = subject
-                };
-                scores.Add(score);
-            }
-            foreach (RecruitmentPlan plan in plans)
-            {
-                SpecialityPriorityVM priority = new()
-                {
-                    PlanId = plan.Id,
-                    NameSpeciality = plan.Speciality.DirectionName ?? plan.Speciality.FullName
-                };
-                priorities.Add(priority);
-            }
-
-            CreateChangeAdmissionVM model = new()
-            {
-                SpecialitiesPriority = priorities,
-                StudentScores = scores,
-                Student = new(),
-                FacultyName = facultyName,
-                GroupId = group.Id
-            };
+            CreateChangeAdmissionVM model = new(scores, priorities);
 
             return View(model);
         }
@@ -108,20 +74,27 @@ namespace DistributionOfStudents.Controllers
 
                 foreach (StudentScore studentScore in model.StudentScores)
                 {
-                    Subject subject = await _subjectsRepository.GetByIdAsync(studentScore.Subject.Id);
-                    studentScore.Subject = subject;
+                    Subject? subject = await _subjectsRepository.GetByIdAsync(studentScore.Subject.Id);
+                    if (subject != null)
+                    {
+                        studentScore.Subject = subject;
+                    }
                 }
                 foreach (SpecialityPriorityVM specialityPriority in model.SpecialitiesPriority.Where(p => p.Priority > 0))
                 {
-                    SpecialityPriority priority = new() { Priority = specialityPriority.Priority, RecruitmentPlan = await _planRepository.GetByIdAsync(specialityPriority.PlanId) };
-                    specialityPriorities.Add(priority);
+                    RecruitmentPlan? plan = await _planRepository.GetByIdAsync(specialityPriority.PlanId);
+                    if (plan != null)
+                    {
+                        SpecialityPriority priority = new() { Priority = specialityPriority.Priority, RecruitmentPlan = plan };
+                        specialityPriorities.Add(priority);
+                    }
                 }
 
                 Admission admission = new()
                 {
-                    GroupOfSpecialties = await _groupsRepository.GetByIdAsync(groupId),
+                    GroupOfSpecialties = await _groupsRepository.GetByIdAsync(groupId) ?? new(),
                     StudentScores = model.StudentScores,
-                    Student = new() { Name = model.Student.Name, Surname = model.Student.Surname, Patronymic = model.Student.Patronymic, GPS = model.Student.GPS },
+                    Student = new() { Name = model.Student.Name.Trim(), Surname = model.Student.Surname.Trim(), Patronymic = model.Student.Patronymic.Trim(), GPS = model.Student.GPS },
                     DateOfApplication = model.DateOfApplication,
                     SpecialityPriorities = specialityPriorities
                 };
@@ -138,44 +111,19 @@ namespace DistributionOfStudents.Controllers
         // GET: Admissions/Edit/5
         public async Task<IActionResult> Edit(string facultyName, int groupId, int id)
         {
-            Admission admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
-            if (admission == null)
+            List<SpecialityPriorityVM> priorities;
+            Admission? admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification().IncludeGroup().IncludeSpecialtyPriorities().IncludeStudentScores());
+            GroupOfSpecialties? group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
+            if (admission == null || group == null)
             {
                 return NotFound();
             }
 
-            GroupOfSpecialties group = await _groupsRepository.GetByIdAsync(groupId, new GroupsOfSpecialitiesSpecification(facultyName).IncludeSpecialties().IncludeSubjects());
-            List<SpecialityPriorityVM> priorities = new();
             List<RecruitmentPlan> plans = await _planRepository.GetAllAsync(new RecruitmentPlansSpecification().WhereFaculty(facultyName).WhereGroup(group));
-            plans = plans.Where(p => group.Specialities.Contains(p.Speciality)).ToList();
+            plans = plans.Where(p => (group.Specialities ?? new()).Contains(p.Speciality)).ToList();
+            priorities = GetSpecialityPrioritiesVM(plans, admission);
 
-            foreach (RecruitmentPlan plan in plans)
-            {
-                SpecialityPriorityVM priority = new()
-                {
-                    PlanId = plan.Id,
-                    NameSpeciality = plan.Speciality.DirectionName ?? plan.Speciality.FullName
-                };
-                foreach (SpecialityPriority admissionPriority in admission.SpecialityPriorities)
-                {
-                    if (plan.Id == admissionPriority.RecruitmentPlan.Id)
-                    {
-                        priority.Priority = admissionPriority.Priority;
-                    }
-                }
-                priorities.Add(priority);
-            }
-
-            CreateChangeAdmissionVM model = new()
-            {
-                Id = id,
-                SpecialitiesPriority = priorities,
-                StudentScores = admission.StudentScores,
-                Student = admission.Student,
-                DateOfApplication = admission.DateOfApplication,
-                FacultyName = facultyName,
-                GroupId = group.Id
-            };
+            CreateChangeAdmissionVM model = new(id, admission, priorities);
 
             return View(model);
         }
@@ -189,7 +137,11 @@ namespace DistributionOfStudents.Controllers
         {
             if (ModelState.IsValid)
             {
-                Admission admission = await _admissionsRepository.GetByIdAsync(model.Id ?? 0, new AdmissionsSpecification().IncludeSpecialtyPriorities().IncludeStudentScores());
+                Admission? admission = await _admissionsRepository.GetByIdAsync(model.Id ?? 0, new AdmissionsSpecification().IncludeSpecialtyPriorities().IncludeStudentScores());
+                if (admission == null)
+                {
+                    return RedirectToAction("Details", "GroupsOfSpecialties", new { facultyName, id = groupId });
+                }
 
                 List<SpecialityPriority> specialityPriorities = new();
                 List<Subject> subjects = new();
@@ -201,20 +153,20 @@ namespace DistributionOfStudents.Controllers
                 foreach (SpecialityPriorityVM specialityPriority in model.SpecialitiesPriority.Where(p => p.Priority > 0))
                 {
                     SpecialityPriority priority = admission.SpecialityPriorities.FirstOrDefault(i => i.RecruitmentPlan.Id == specialityPriority.PlanId) ??
-                        new() { RecruitmentPlan = await _planRepository.GetByIdAsync(specialityPriority.PlanId) };
+                        new() { RecruitmentPlan = await _planRepository.GetByIdAsync(specialityPriority.PlanId) ?? new() };
                     priority.Priority = specialityPriority.Priority;
                     specialityPriorities.Add(priority);
                 }
 
                 admission.SpecialityPriorities = specialityPriorities;
-                admission.Student.Surname = model.Student.Surname;
-                admission.Student.Name = model.Student.Name;
-                admission.Student.Patronymic = model.Student.Patronymic;
+                admission.Student.Surname = model.Student.Surname.Trim();
+                admission.Student.Name = model.Student.Name.Trim();
+                admission.Student.Patronymic = model.Student.Patronymic.Trim();
                 admission.Student.GPS = model.Student.GPS;
 
                 await _admissionsRepository.UpdateAsync(admission);
 
-                _logger.LogInformation("Заявка студента - {Surname} {Name} {Patronymic} - изменена", admission.Student.Surname, admission.Student.Name, admission.Student.Patronymic);
+                _logger.LogInformation("Заявка абитуриента - {Surname} {Name} {Patronymic} - изменена", admission.Student.Surname, admission.Student.Name, admission.Student.Patronymic);
                 try
                 {
                     await _admissionsRepository.UpdateAsync(admission);
@@ -237,9 +189,12 @@ namespace DistributionOfStudents.Controllers
 
         public async Task<IActionResult> Delete(string facultyName, int groupId, int id)
         {
-            Admission admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification());
-            await _admissionsRepository.DeleteAsync(id);
-            _logger.LogInformation("Заявка студента - {Surname} {Name} {Patronymic} - была удалена", admission.Student.Surname, admission.Student.Name, admission.Student.Patronymic);
+            Admission? admission = await _admissionsRepository.GetByIdAsync(id, new AdmissionsSpecification());
+            if (admission != null)
+            {
+                await _admissionsRepository.DeleteAsync(id);
+                _logger.LogInformation("Заявка студента - {Surname} {Name} {Patronymic} - была удалена", admission.Student.Surname, admission.Student.Name, admission.Student.Patronymic);
+            }
 
             return RedirectToAction("Details", "GroupsOfSpecialties", new { facultyName, id = groupId });
         }
@@ -248,6 +203,25 @@ namespace DistributionOfStudents.Controllers
         {
             var admissions = await _admissionsRepository.GetAllAsync();
             return admissions.Any(e => e.Id == id);
+        }
+
+        private static List<SpecialityPriorityVM> GetSpecialityPrioritiesVM(List<RecruitmentPlan> plans, Admission admission)
+        {
+            List<SpecialityPriorityVM> priorities = new();
+            foreach (RecruitmentPlan plan in plans)
+            {
+                SpecialityPriorityVM priority = new(plan);
+                foreach (SpecialityPriority admissionPriority in admission.SpecialityPriorities)
+                {
+                    if (plan.Id == admissionPriority.RecruitmentPlan.Id)
+                    {
+                        priority.Priority = admissionPriority.Priority;
+                    }
+                }
+                priorities.Add(priority);
+            }
+
+            return priorities;
         }
     }
 }
