+++ DistributionOfStudents/Controllers/SpecialitiesController.cs
@@ -1,15 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
-using Microsoft.EntityFrameworkCore;
-using DistributionOfStudents.Data;
+﻿using DistributionOfStudents.Data.Interfaces;
 using DistributionOfStudents.Data.Models;
-using DistributionOfStudents.Data.Interfaces;
 using DistributionOfStudents.Data.Specifications;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.CodeAnalysis;
+using Microsoft.EntityFrameworkCore;
 
 namespace DistributionOfStudents.Controllers
 {
@@ -30,8 +24,8 @@ namespace DistributionOfStudents.Controllers
         // GET: Specialties/Create
         public async Task<IActionResult> Create(string facultyName)
         {
-            Faculty faculty = await _facultiesRepository.GetByShortNameAsync(facultyName, new FacultiesSpecification());
-            if(faculty == null)
+            Faculty? faculty = await _facultiesRepository.GetByShortNameAsync(facultyName, new FacultiesSpecification());
+            if (faculty == null)
             {
                 return NotFound();
             }
@@ -53,7 +47,7 @@ namespace DistributionOfStudents.Controllers
         {
             if (ModelState.IsValid)
             {
-                Faculty faculty = await _facultiesRepository.GetByShortNameAsync(specialty.Faculty.ShortName, new FacultiesSpecification());
+                Faculty? faculty = await _facultiesRepository.GetByShortNameAsync(specialty.Faculty.ShortName, new FacultiesSpecification());
                 if (faculty == null)
                 {
                     return NotFound();
@@ -70,8 +64,8 @@ namespace DistributionOfStudents.Controllers
         // GET: Specialties/Edit/5
         public async Task<IActionResult> Edit(string facultyName, int id)
         {
-            Speciality specialty = await _specialtiesRepository.GetByIdAsync(id);
-            Faculty faculty = await _facultiesRepository.GetByShortNameAsync(facultyName, new FacultiesSpecification());
+            Speciality? specialty = await _specialtiesRepository.GetByIdAsync(id);
+            Faculty? faculty = await _facultiesRepository.GetByShortNameAsync(facultyName, new FacultiesSpecification());
             if (specialty == null || faculty == null)
             {
                 return NotFound();
@@ -118,9 +112,12 @@ namespace DistributionOfStudents.Controllers
         // GET: FacultiesController/Delete/5
         public async Task<RedirectToActionResult> Delete(string facultyName, int id)
         {
-            Speciality specialty = await _specialtiesRepository.GetByIdAsync(id);
-            await _specialtiesRepository.DeleteAsync(id);
-            _logger.LogInformation("Специальность - {SpecialityName} - была удалена", specialty.FullName);
+            Speciality? specialty = await _specialtiesRepository.GetByIdAsync(id);
+            if (specialty != null)
+            {
+                await _specialtiesRepository.DeleteAsync(id);
+                _logger.LogInformation("Специальность - {SpecialityName} - была удалена", specialty.FullName);
+            }
 
             return RedirectToAction("Details", "Faculties", new { name = facultyName, });
         }
